Handle MUC message handling and reflections.
If a MUC message is sent like a regular chat msg, persistence
will get confused when the server reflect the message and create
a duplicate.

Finding "self sent" message is a bit difficult, MessageHandler
has no reference to MucManager, so let's store the message as if
directly received from groupchat, that way reflection will be found
in persistence and not added.

Issue is that own messages can't be displayed differently in
groupchat, but as other clients self-message were not, it's
actually making things a bit more consistent.